moved all the entities back
