+++ libs/common/src/entities/index.ts
@@ -1,4 +1,1 @@
 export * from './user.entity';
-export * from './chat.entity';
-export * from './message.entity';
-export * from './guide.entity';
