+++ src/chat/dto/update-chat.dto.ts
@@ -6,7 +6,7 @@ import {
   IsOptional,
   ValidateNested,
 } from 'class-validator';
-import { Message } from '@app/common';
+import { Message } from '../message/entities/message.entity';
 import { Type } from 'class-transformer';
 
 export class UpdateChatDto {
