+++ src/chat/entities/chat.entity.ts
@@ -6,8 +6,8 @@ import {
   OneToMany,
   PrimaryGeneratedColumn,
 } from 'typeorm';
-import { Message } from './message.entity';
-import { User } from './user.entity';
+import { Message } from '../message/entities/message.entity';
+import { User } from '@app/common';
 @Entity('Chat')
 export class Chat {
   @PrimaryGeneratedColumn()
