+++ src/chat/message/message.service.ts
@@ -9,7 +9,7 @@ import { CreateMessageDto } from './dto/create-message.dto';
 import { UpdateMessageDto } from './dto/update-message.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Message } from '../../../libs/common/src/entities/message.entity';
+import { Message } from './entities/message.entity';
 import { CHAT_ERROR_MESSAGES } from '@app/common';
 
 @Injectable()
