+++ src/guide/entities/guide.entity.ts
@@ -6,7 +6,7 @@ import {
   ManyToOne,
   PrimaryGeneratedColumn,
 } from 'typeorm';
-import { User } from './user.entity';
+import { User } from '@app/common';
 
 @Entity('Guide')
 export class Guide {
